packages/engine/pyproject.toml
[project]
name = "polypaper-engine"
version = "1.0.0"
description = "Polypaper Strategy Engine"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary>=2.9.9",
    "redis>=5.0.1",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
